refactor: adopt DDD module structure
