--- src/modules/advertising/route.ts
@@ -1,9 +1,0 @@
-import { Router } from 'express';
-import { advertisingController } from '@/modules/advertising/controller/controller';
-import asyncHandler from '@/shared/utils/asyncHandler';
-
-const router = Router();
-
-router.get('/sync', asyncHandler(advertisingController.sync));
-
-export default router;
