+++ src/modules/analytics/application/dto/margin.dto.ts
