+++ src/modules/posting/application/dto/api-posting.dto.ts
