+++ src/modules/posting/application/dto/posting.dto.ts
