+++ src/modules/posting/infrastructure/posting-item.api.ts
@@ -1,0 +1,18 @@
+import {sellerClient} from "@/infrastructure/clients/ozon/seller";
+import {ApiPostingDto} from "@/modules/posting/application/dto/api-posting.dto";
+
+export async function postingsFetchItem(postingNumber: string): Promise<ApiPostingDto> {
+    const response = await sellerClient.post(
+        'v2/posting/fbo/get',
+        {
+            posting_number: postingNumber,
+            with: {
+                analytics_data: true,
+                financial_data: true,
+                legal_info: true,
+            },
+        },
+    );
+
+    return response.data.result;
+}
