--- src/modules/unit/route.ts
@@ -1,9 +1,0 @@
-import { Router } from 'express';
-import { unitController } from '@/modules/unit/controller/controller';
-import asyncHandler from '@/shared/utils/asyncHandler';
-
-const router = Router();
-
-router.get('/sync', asyncHandler(unitController.sync));
-
-export default router;
